+++ intellij-plugin-verifier/verifier-intellij/src/main/kotlin/com/jetbrains/pluginverifier/response/VerifierServiceApi.kt
@@ -18,7 +18,6 @@ enum class VerificationResultTypeDto {
   PROBLEMS,
   INVALID_PLUGIN,
   NON_DOWNLOADABLE,
-  UNCHANGED,
   CRITICAL;
 }
